Add files via upload
Background
Non-coding RNAs (ncRNAs) have a multitude of roles in the cell, many of which remain to be discovered. However, it is difficult to detect novel ncRNAs in biochemical screens. To advance biological knowledge, machine learning methods that can accurately detect ncRNAs in sequenced genomes are therefore desirable. In this assignment, you will be exploring the data with structural information of the RNA molecules to understand and evaluate 6 pre-trained classifiers stored in your project directory “models/” as “Model_1.pkl”, “Model_2.pkl”, “Model_3.pkl”, “Model_4.pkl”, “Model_5.pkl”, “Model_6.pkl”.